move timeshift code
